refactor(#3036): move CB and import to modules.fim
