+++ deps/wazuh_testing/wazuh_testing/modules/fim/__init__.py
@@ -10,6 +10,7 @@ maintain if one of them changes in the future.
 import sys
 
 # Variables
+SIZE_LIMIT_CONFIGURED_VALUE = 10240
 
 # Key variables
 MONITORED_KEY = 'SOFTWARE\\random_key'
@@ -55,15 +56,15 @@ WHODATA_MODE = 'whodata'
 
 # Yaml Configuration
 YAML_CONF_REGISTRY_RESPONSE = 'wazuh_conf_registry_responses_win32.yaml'
-
 YAML_CONF_SYNC_WIN32 = 'wazuh_sync_conf_win32.yaml'
 YAML_CONF_DIFF = 'wazuh_conf_diff.yaml'
 YAML_CONF_MAX_EPS_SYNC = 'wazuh_sync_conf_max_eps.yaml'
 
+
 # Synchronization options
 SYNCHRONIZATION_ENABLED = 'SYNCHRONIZATION_ENABLED'
-
 SYNCHRONIZATION_REGISTRY_ENABLED = 'SYNCHRONIZATION_REGISTRY_ENABLED'
+
 
 # Callbacks message
 CB_INTEGRITY_CONTROL_MESSAGE = r".*Sending integrity control message: (.+)$"
@@ -71,17 +72,14 @@ CB_MAXIMUM_FILE_SIZE = r".*Maximum file size limit to generate diff information 
 CB_AGENT_CONNECT = r".* Connected to the server .*"
 CB_INODE_ENTRIES_PATH_COUNT = r".*Fim inode entries: '(\d+)', path count: '(\d+)'"
 CB_DETECT_FIM_EVENT = r".*Sending FIM event: (.+)$"
-
 CB_DATABASE_FULL_COULD_NOT_INSERT_VALUE = r".*registry_value.*Couldn't insert ('.*') entry into DB. The DB is full.*"
 CB_DATABASE_FULL_COULD_NOT_INSERT_KEY = r".*registry_key.*Couldn't insert ('.*')? entry into DB\. The DB is full.*"
-
 CB_COUNT_REGISTRY_ENTRIES = r".*Fim registry entries count: '(\d+)'"
 CB_COUNT_REGISTRY_VALUE_ENTRIES = r".*Fim registry values entries count: '(\d+)'"
 CB_REGISTRY_DBSYNC_NO_DATA = r".*#!-fim_registry_(.*) dbsync no_data (.+)"
 CB_REGISTRY_LIMIT_CAPACITY = r".*Registry database is (\d+)% full."
 CB_REGISTRY_DB_BACK_TO_NORMAL = r".*(The registry database status returns to normal)."
 CB_REGISTRY_LIMIT_VALUE = r".*Maximum number of registry values to be monitored: '(\d+)'"
-
 CB_FILE_LIMIT_CAPACITY = r".*File database is (\d+)% full."
 CB_FILE_LIMIT_BACK_TO_NORMAL = r".*(Sending DB back to normal alert)."
 CB_FIM_ENTRIES_COUNT = r".*Fim file entries count: '(\d+)'"
@@ -90,12 +88,14 @@ CB_FILE_LIMIT_DISABLED = r".*(No limit set) to maximum number of file entries to
 CB_PATH_MONITORED_REALTIME = r".*Directory added for real time monitoring: (.*)"
 CB_PATH_MONITORED_WHODATA = r".*Added audit rule for monitoring directory: (.*)"
 CB_PATH_MONITORED_WHODATA_WINDOWS = r".*Setting up SACL for (.*)"
+CB_IGNORING_DUE_TO_SREGEX = r".*?Ignoring path '(.*)' due to sregex '(.*)'.*"
+CB_IGNORING_DUE_TO_PATTERN = r".*?Ignoring path '(.*)' due to pattern '(.*)'.*"
+
 
 # Error message
 
 ERR_MSG_AGENT_DISCONNECT = 'Agent couldn\'t connect to server.'
 ERR_MSG_INTEGRITY_CONTROL_MSG = 'Didn\'t receive control message(integrity_check_global)'
-
 ERR_MSG_DATABASE_PERCENTAGE_FULL_ALERT = 'Did not receive expected "DEBUG: ...: database is ...% full" alert'
 ERR_MSG_WRONG_CAPACITY_LOG_DB_LIMIT = 'Wrong capacity log for DB file_limit'
 ERR_MSG_DB_BACK_TO_NORMAL = 'Did not receive expected "DEBUG: ... database status returns to normal." event'
@@ -104,25 +104,20 @@ ERR_MSG_WRONG_VALUE_FOR_DATABASE_FULL = 'Wrong value for full database alert.'
 ERR_MSG_DATABASE_FULL_COULD_NOT_INSERT = 'Did not receive expected "DEBUG: ...: Couldn\'t insert \'...\' entry \
                                           into DB. The DB is full, ..." event'
 ERR_MSG_DATABASE_FULL_ALERT_EVENT = 'Did not receive expected "DEBUG: ...: Sending DB 100% full alert." event'
-
 ERR_MSG_WRONG_NUMBER_OF_ENTRIES = 'Wrong number of entries counted.'
 ERR_MSG_WRONG_INODE_PATH_COUNT = 'Wrong number of inodes and path count'
 ERR_MSG_FIM_INODE_ENTRIES = 'Did not receive expected "Fim inode entries: ..., path count: ..." event'
-
 ERR_MSG_FIM_REGISTRY_ENTRIES = 'Did not receive expected "Fim Registry entries count: ..." event'
 ERR_MSG_FIM_REGISTRY_VALUE_ENTRIES = 'Did not receive expected "Fim Registry value entries count: ..." event'
-
 ERR_MSG_REGISTRY_LIMIT_VALUES = 'Did not receive expected "DEBUG: ...: Maximum number of registry values to \
                                  be monitored: ..." event'
 ERR_MSG_WRONG_REGISTRY_LIMIT_VALUE = 'Wrong value for db_value_limit registries tag.'
-
 ERR_MSG_FILE_LIMIT_VALUES = 'Did not receive expected "DEBUG: ...: Maximum number of entries to be monitored: \
                              ..." event'
 ERR_MSG_WRONG_FILE_LIMIT_VALUE = 'Wrong value for file_limit.'
 ERR_MSG_FILE_LIMIT_DISABLED = 'Did not receive expected "DEBUG: ...: No limit set to maximum number of entries \
                                to be monitored" event'
 ERR_MSG_MAXIMUM_FILE_SIZE = 'Did not receive expected "Maximum file size limit configured to \'... KB\'..." event'
-
 ERR_MSG_NO_EVENTS_EXPECTED = 'No events should be detected.'
 ERR_MSG_DELETED_EVENT_NOT_RECIEVED = 'Did not receive expected deleted event'
 ERR_MSG_FIM_EVENT_NOT_RECIEVED = 'Did not receive expected "Sending FIM event: ..." event'
@@ -133,6 +128,9 @@ ERR_MSG_WRONG_VALUE_MAXIMUM_FILE_SIZE = 'Wrong value for diff_size_limit'
 ERR_MSG_INTEGRITY_OR_WHODATA_NOT_STARTED = 'Did not receive expected "File integrity monitoring real-time Whodata \
                                             engine started" or "Initializing FIM Integrity Synchronization check"'
 ERR_MSG_INTEGRITY_CHECK_EVENT = 'Did not receive expected "Initializing FIM Integrity Synchronization check" event'
+ERR_MSG_CONTENT_CHANGES_EMPTY = "content_changes is empty"
+ERR_MSG_CONTENT_CHANGES_NOT_EMPTY = "content_changes isn't empty"
+
 
 # Setting Local_internal_option file
 if sys.platform == 'win32':
