+++ tests/integration/test_fim/test_files/test_basic_usage/test_basic_usage_entries_match_path_count.py
@@ -75,7 +75,7 @@ from wazuh_testing.fim import LOG_FILE_PATH, generate_params, create_file, REGUL
 from wazuh_testing.tools import PREFIX
 from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
 from wazuh_testing.tools.monitoring import FileMonitor
-from wazuh_testing.fim_module.event_monitor import callback_entries_path_count
+from wazuh_testing.modules.fim.event_monitor import callback_entries_path_count
 
 # Marks
